Addition of the report_card.py functionality
